change all bytes to represent intf to int (Comment: #8788)
git-svn-id: https://ut.svn.beanstalkapp.com/ngfw/work/src@29042 8ebf8b4e-06f3-0310-961f-89f74618be54

+++ openvpn/api/com/untangle/node/openvpn/VpnGroup.java
@@ -43,7 +43,7 @@ public class VpnGroup extends Rule implements Validatable
 {
 
     /* The interface that clients from the client pool are associated with */
-    private byte intf;
+    private int intf;
 
     private IPAddress address;
     private IPAddress netmask;
@@ -105,12 +105,12 @@ public class VpnGroup extends Rule implements Validatable
      * @return Default interface to associate VPN traffic with.
      */
     @Transient
-    public byte getIntf()
+    public int getIntf()
     {
         return this.intf;
     }
 
-    public void setIntf( byte intf )
+    public void setIntf( int intf )
     {
         this.intf = intf;
     }
